Move the settings panel into its own module
Third and last slice of #67. settings-panel.js owns the gear, the panel and
everything drawn in it, and is the composition root for the group: it builds
theme.js and settings-storage.js itself, so navio.js constructs one thing.
The graph runs one way - navio.js -> settings-panel -> {settings-storage,
theme} - and applyTheme lives with the panel rather than the theme table
because it restyles the panel's own elements.

Five bindings leave the closure's let chain: settingsButton, settingsPanel,
collapsedSections, panelPointerHeld and liftedClips.

Two things navio.js used to reach into directly, and now asks for:

- The layout pass positioned the gear by touching settingsButton, 2,000 lines
  from anything else that knew about it. It calls positionButton() now.
- destroy() detached the light-dismiss listener itself. That listener is
  registered on document in the CAPTURE phase, so removal needs the identical
  function; registration and removal both live in the module now and navio.js
  calls settings.destroy().

The panel is constructed where the slice was, not at the top of the closure.
It registers seven public methods on nv, and OPTION_NAMES snapshots
Object.keys(nv) at ~336 - built any earlier, getOptions() starts reporting
those methods as options and 104-describe.spec.js fails. npm run check does not
run e2e, so that one would have landed green.

Three defects the mechanical rewrite produced, all caught before running:

- `instanceId,` is an object-literal SHORTHAND where the picker's options are
  built; rewriting it to `ctx.instanceId,` is a syntax error. It reads
  `instanceId: ctx.instanceId` now. The same class as the `height:` key bug in
  the previous commit.
- `${visibleAttribs().length}` sits inside a template literal. The rewriter
  masks strings so it will not corrupt getSettingsCode's generated output, and
  that masking hid a real ${} interpolation. eslint's no-undef caught it.
- The extracted range ran one line too far and swallowed isVertical's JSDoc.

instanceId is deliberately NOT rewritten inside defaultAttribPicker, which
declares a parameter of the same name that shadows the closure const.

navio.js: 5,799 -> 4,545 lines. Gate: 102 unit tests, 286 e2e,
docs/ai/API.md unchanged.
